created new TrendingHashtags component
